+++ CHANGELOG.md
@@ -62,6 +62,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix duplicated labels in column picker - #1197 by @orzechdev
 - Fix forbidden null sending as attribute value - #1201 by @orzechdev
 - Fix missing call for update metadata mutation - #1207 by @orzechdev
+- Fix order links on home page - #1219 by @jwm0
 
 # 2.11.1